+++ src/Telegram.Bot/Converters/BanTimeUnixDateTimeConverter.cs
@@ -6,23 +6,18 @@ namespace Telegram.Bot.Converters
 {
     internal class BanTimeUnixDateTimeConverter : UnixDateTimeConverter
     {
-        static readonly DateTime DefaultUtc = DateTime.SpecifyKind(default, DateTimeKind.Utc);
-
         public override object? ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
         {
             var nonNullable = Nullable.GetUnderlyingType(objectType) is null;
 
-            if (reader.TokenType == JsonToken.Integer && reader.Value is 0L)
-            {
-                return nonNullable ? DefaultUtc : null;
-            }
-
-            return base.ReadJson(reader, objectType, existingValue, serializer);
+            return reader.TokenType == JsonToken.Integer && reader.Value is 0L
+                ? nonNullable ? default : null
+                : base.ReadJson(reader, objectType, existingValue, serializer);
         }
 
         public override void WriteJson(JsonWriter writer, object? value, JsonSerializer serializer)
         {
-            if (value is null || value.Equals(DefaultUtc))
+            if (value is null || value.Equals(default(DateTime)))
             {
                 writer.WriteValue(0);
             }