+++ src/Telegram.Bot/Types/InlineQueryResults/InlineQueryResult/Documentation.cs
@@ -1,61 +1,59 @@
 ﻿#nullable disable
 #pragma warning disable 169
+#pragma warning disable CA1823
 // ReSharper disable InconsistentNaming
 // ReSharper disable once CheckNamespace
 namespace Telegram.Bot.Types.InlineQueryResults
 {
     // ReSharper disable once UnusedType.Global
-    internal abstract class Documentation
+    internal static class Documentation
     {
-        Documentation()
-        { }
-
         /// <summary>
         /// Content of the message to be sent instead of the result
         /// </summary>
-        object InputMessageContent { get; set; }
+        static object InputMessageContent;
 
         /// <summary>
         /// Caption of the result to be sent, 0-1024 characters after entities parsing
         /// </summary>
-        object Caption;
+        static object Caption;
 
         /// <summary>
         /// Mode for parsing entities in the result caption. See
         /// <see href="https://core.telegram.org/bots/api#formatting-options">formatting options</see>
         /// for more details.
         /// </summary>
-        object ParseMode;
+        static object ParseMode;
 
         /// <summary>
         /// List of special entities that appear in the caption, which can be specified
         /// instead of <see cref="ParseMode"/>
         /// </summary>
-        object CaptionEntities;
+        static object CaptionEntities;
 
         /// <summary>
         /// Location latitude in degrees
         /// </summary>
-        object Latitude;
+        static object Latitude;
 
         /// <summary>
         /// Location longitude in degrees
         /// </summary>
-        object Longitude;
+        static object Longitude;
 
         /// <summary>
         /// Thumbnail width
         /// </summary>
-        object ThumbWidth;
+        static object ThumbWidth;
 
         /// <summary>
         /// Thumbnail height
         /// </summary>
-        object ThumbHeight;
+        static object ThumbHeight;
 
         /// <summary>
         /// Url of the thumbnail for the result
         /// </summary>
-        object ThumbUrl;
+        static object ThumbUrl;
     }
 }