+++ src/Telegram.Bot/Types/InputFiles/IInputMediaThumb.cs
@@ -8,7 +8,9 @@ namespace Telegram.Bot.Types
     public interface IInputMediaThumb
     {
         /// <summary>
-        /// Optional. Thumbnail of the file sent; can be ignored if thumbnail generation for the file is supported server-side. The thumbnail should be in JPEG format and less than 200 kB in size. A thumbnail's width and height should not exceed 320. Ignored if the file is not uploaded using multipart/form-data.
+        /// Optional. Thumbnail of the file sent; can be ignored if thumbnail generation for the file is supported
+        /// server-side. The thumbnail should be in JPEG format and less than 200 kB in size. A thumbnail's width and
+        /// height should not exceed 320. Ignored if the file is not uploaded using multipart/form-data.
         /// </summary>
         InputMedia? Thumb { get; }
     }