+++ test/Telegram.Bot.Tests.Unit/Serialization/ChatMemberSerializationTests.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using System;
+using Newtonsoft.Json;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
 using Xunit;
@@ -72,6 +73,154 @@ namespace Telegram.Bot.Tests.Unit.Serialization
             Assert.Contains(@"""last_name"":""Last Name""", chatMemberJson);
             Assert.Contains(@"""username"":""test_bot""", chatMemberJson);
             Assert.Contains(@"""language_code"":""en_US""", chatMemberJson);
+        }
+
+        [Fact]
+        public void Should_Serialize_Chat_Member_Banned()
+        {
+            var creator = new ChatMemberBanned()
+            {
+                User = new User
+                {
+                    Id = 12345,
+                    IsBot = true,
+                    FirstName = "First Name",
+                    LastName = "Last Name",
+                    Username = "test_bot",
+                    LanguageCode = "en_US",
+                },
+                UntilDate = new DateTime(2021, 4, 2, 0, 0, 0, DateTimeKind.Utc)
+            };
+
+            var chatMemberJson = JsonConvert.SerializeObject(creator);
+
+            Assert.Contains(@"""until_date"":1617321600", chatMemberJson);
+            Assert.Contains(@"""status"":""kicked""", chatMemberJson);
+            Assert.Contains(@"""user"":{", chatMemberJson);
+            Assert.Contains(@"""id"":12345", chatMemberJson);
+            Assert.Contains(@"""is_bot"":true", chatMemberJson);
+            Assert.Contains(@"""first_name"":""First Name""", chatMemberJson);
+            Assert.Contains(@"""last_name"":""Last Name""", chatMemberJson);
+            Assert.Contains(@"""username"":""test_bot""", chatMemberJson);
+            Assert.Contains(@"""language_code"":""en_US""", chatMemberJson);
+        }
+
+        [Fact]
+        public void Should_Serialize_Chat_Member_Banned_2()
+        {
+            var creator = new ChatMemberBanned()
+            {
+                User = new User
+                {
+                    Id = 12345,
+                    IsBot = true,
+                    FirstName = "First Name",
+                    LastName = "Last Name",
+                    Username = "test_bot",
+                    LanguageCode = "en_US",
+                },
+            };
+
+            var chatMemberJson = JsonConvert.SerializeObject(creator);
+
+            Assert.DoesNotContain(@"""until_date""", chatMemberJson);
+            Assert.Contains(@"""status"":""kicked""", chatMemberJson);
+            Assert.Contains(@"""user"":{", chatMemberJson);
+            Assert.Contains(@"""id"":12345", chatMemberJson);
+            Assert.Contains(@"""is_bot"":true", chatMemberJson);
+            Assert.Contains(@"""first_name"":""First Name""", chatMemberJson);
+            Assert.Contains(@"""last_name"":""Last Name""", chatMemberJson);
+            Assert.Contains(@"""username"":""test_bot""", chatMemberJson);
+            Assert.Contains(@"""language_code"":""en_US""", chatMemberJson);
+        }
+
+        [Fact]
+        public void Should_Deserialize_Chat_Member_Banned()
+        {
+            var json = @"
+                {
+                    ""status"": ""kicked"",
+                    ""until_date"": 1617321600,
+                    ""user"": {
+                        ""id"": 12345,
+                        ""is_bot"": true,
+                        ""first_name"": ""First Name"",
+                        ""last_name"": ""Last Name"",
+                        ""username"": ""test_bot"",
+                        ""language_code"": ""en_US""
+                    }
+                }";
+
+            var bannedUser = JsonConvert.DeserializeObject<ChatMemberBanned>(json);
+
+            Assert.Equal(ChatMemberStatus.Kicked, bannedUser.Status);
+            Assert.Equal(new DateTime(2021, 4, 2, 0, 0, 0, DateTimeKind.Utc), bannedUser.UntilDate);
+            Assert.NotNull(bannedUser.User);
+            Assert.Equal(12345, bannedUser.User.Id);
+            Assert.True(bannedUser.User.IsBot);
+            Assert.Equal("First Name", bannedUser.User.FirstName);
+            Assert.Equal("Last Name", bannedUser.User.LastName);
+            Assert.Equal("test_bot", bannedUser.User.Username);
+            Assert.Equal("en_US", bannedUser.User.LanguageCode);
+        }
+
+        [Fact]
+        public void Should_Deserialize_Chat_Member_Banned_2()
+        {
+            var json = @"
+                {
+                    ""status"": ""kicked"",
+                    ""until_date"": 0,
+                    ""user"": {
+                        ""id"": 12345,
+                        ""is_bot"": true,
+                        ""first_name"": ""First Name"",
+                        ""last_name"": ""Last Name"",
+                        ""username"": ""test_bot"",
+                        ""language_code"": ""en_US""
+                    }
+                }";
+
+            var bannedUser = JsonConvert.DeserializeObject<ChatMemberBanned>(json);
+
+            Assert.Equal(ChatMemberStatus.Kicked, bannedUser.Status);
+            Assert.Null(bannedUser.UntilDate);
+            Assert.NotNull(bannedUser.User);
+            Assert.Equal(12345, bannedUser.User.Id);
+            Assert.True(bannedUser.User.IsBot);
+            Assert.Equal("First Name", bannedUser.User.FirstName);
+            Assert.Equal("Last Name", bannedUser.User.LastName);
+            Assert.Equal("test_bot", bannedUser.User.Username);
+            Assert.Equal("en_US", bannedUser.User.LanguageCode);
+        }
+
+        [Fact]
+        public void Should_Deserialize_Chat_Member_Banned_3()
+        {
+            var json = @"
+                {
+                    ""status"": ""kicked"",
+                    ""user"": {
+                        ""id"": 12345,
+                        ""is_bot"": true,
+                        ""first_name"": ""First Name"",
+                        ""last_name"": ""Last Name"",
+                        ""username"": ""test_bot"",
+                        ""language_code"": ""en_US""
+                    }
+                }";
+
+            var bannedUser = JsonConvert.DeserializeObject<ChatMemberBanned>(json);
+
+            Assert.Equal(ChatMemberStatus.Kicked, bannedUser.Status);
+            Assert.Null(bannedUser.UntilDate);
+            Assert.NotNull(bannedUser.User);
+            Assert.Equal(12345, bannedUser.User.Id);
+            Assert.True(bannedUser.User.IsBot);
+            Assert.Equal("First Name", bannedUser.User.FirstName);
+            Assert.Equal("Last Name", bannedUser.User.LastName);
+            Assert.Equal("test_bot", bannedUser.User.Username);
+            Assert.Equal("en_US", bannedUser.User.LanguageCode);
         }
     }
 }